+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [unreleased]
 
+### Changed
+
+- Debugger port of headless browser changed from 9222 to 9223.
+
 ### Fixed
 
 - Detection of dependencies does not break build in second rebuild. It also does not show as missing dependency when it is in devDependencies.