+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### Changed
 
+- By default all @types modules will be now referenced by TypeScript.
 - Debugger port of headless browser changed from 9222 to 9223.
 
 ### Fixed
 
-- Detection of dependencies does not break build in second rebuild. It also does not show as missing dependency when it is in devDependencies.
+- Detection of dependencies does not break build in second rebuild. It also does not show as missing dependency when it is in devDependencies. Automatically ignore @types/ dependencies.
 
 ## 0.56.0